requirements.txt
django-widget-tweaks
django-haystack
elasticsearch
django-form-utils
Pillow
django-floppyforms
celery
django-celery